[FIX] z_number and linear_expression constructors took integer parameters. If the parameter was not in the range [INT_MIN, INT_MAX] then signed integer overflow was occurring. The parameters are now of type signed long long int.

+++ include/crab/domains/linear_constraints.hpp
@@ -137,7 +137,7 @@ namespace ikos {
     linear_expression(Number n): 
         _map(map_ptr(new map_t)), _cst(n) { }
     
-    linear_expression(int n): 
+    linear_expression(signed long long int n): 
         _map(map_ptr(new map_t)), _cst(Number(n)) { }
     
     linear_expression(variable_t x):
